feat(compat): implement INI and conf translation layer
- Add src/compat/ini/ module for Python Supervisor .conf/.ini parsing, normalization, and adapter mapping
- Support loose booleans, autorestart mapping, environment parser, signal aliases, umask, and log path handling
- Implement Python ConfigParser parsing with relative-indent line continuation, inline comment stripping, and %(here)s expansion
- Support [include] glob file merging with circular include prevention
- Update SupervisorConfig::from_file to dispatch .yaml/.yml to YAML and all other files to INI loader
- Relax program and group priority range to [0, 999] across config schema and Program traits
- Preserve unexpanded literal  and  shell expressions in StringExpression
- Add integration and equivalence tests in tests/ini_tests.rs

+++ src/program/process.rs
@@ -180,7 +180,7 @@ impl Program for ProcessProgram {
         &self.config.name
     }
 
-    fn priority(&self) -> u8 {
+    fn priority(&self) -> u32 {
         self.config.priority
     }
